Lag egen service for Endringstidspunkt for å unngå sirkelavhengighet
