+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/BeregningService.kt
@@ -1,10 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.beregning
 
-import no.nav.familie.ba.sak.common.TIDENES_ENDE
-import no.nav.familie.ba.sak.common.TIDENES_MORGEN
 import no.nav.familie.ba.sak.common.toYearMonth
 import no.nav.familie.ba.sak.kjerne.behandling.Behandlingutils
-import no.nav.familie.ba.sak.kjerne.behandling.Behandlingutils.hentSisteBehandlingSomErIverksatt
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingRepository
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
@@ -222,25 +219,5 @@ class BeregningService(
             this.endretDato = LocalDate.now()
             this.opphørFom = opphørsdato?.toYearMonth()
         }
-    }
-
-    fun finnEndringstidpunkForBehandling(behandlingId: Long): LocalDate {
-        val nyBehandling = behandlingRepository.finnBehandling(behandlingId)
-
-        val iverksatteBehandlinger = behandlingRepository.finnIverksatteBehandlinger(fagsakId = nyBehandling.fagsak.id)
-        val gammelBehandling = hentSisteBehandlingSomErIverksatt(iverksatteBehandlinger)
-            ?: return TIDENES_MORGEN
-
-        val nyeAndelerTilkjentYtelse = andelTilkjentYtelseRepository
-            .finnAndelerTilkjentYtelseForBehandling(behandlingId = behandlingId)
-
-        val gamleAndelerTilkjentYtelse = andelTilkjentYtelseRepository
-            .finnAndelerTilkjentYtelseForBehandling(behandlingId = gammelBehandling.id)
-
-        val perioderMedEndringer = nyeAndelerTilkjentYtelse.hentPerioderMedEndringerFra(
-            forrigeAndelerTilkjentYtelse = gamleAndelerTilkjentYtelse
-        )
-
-        return perioderMedEndringer.minOfOrNull { it.fom } ?: TIDENES_ENDE
     }
 }
